feat: better support of Object.hasOwn and "in"
Prior to this commit `Object.hasOwn` and `in` operator would mark object
keys as tracked meaning that different values for the key would result
in object being reported as changed even though the actual value was
never read.

With this commit we start tracking `hasKey` and `hasOwnKey` separately
from regular properties and check for the key presence/absence in
`isChanged`.